@@ -33,14 +33,14 @@ class ControllerPathValidation implements IValidation
 			}
 
 			// MUST: Starts with slash (/)
-			if (substr($path, 0, 1) !== '/') {
+			if (!str_starts_with($path, '/')) {
 				throw new InvalidSchemaException(
 					sprintf('@Path "%s" in "%s" must starts with "/" (slash).', $path, $controller->getClass())
 				);
 			}
 
 			// MUST NOT: Ends with slash (/)
-			if (substr($path, -1, 1) === '/') {
+			if (str_ends_with($path, '/')) {
 				throw new InvalidSchemaException(
 					sprintf('@Path "%s" in "%s" must not ends with "/" (slash).', $path, $controller->getClass())
 				);
