@@ -19,6 +19,7 @@ class PathValidation implements IValidation
 	protected function validateRequirements(SchemaBuilder $builder): void
 	{
 		$controllers = $builder->getControllers();
+
 		foreach ($controllers as $controller) {
 			foreach ($controller->getMethods() as $method) {
 				if ($method->getPath() === '') {
@@ -45,7 +46,7 @@ class PathValidation implements IValidation
 				$path = $method->getPath();
 
 				// MUST: Starts with slash (/)
-				if (substr($path, 0, 1) !== '/') {
+				if (!str_starts_with($path, '/')) {
 					throw (new InvalidSchemaException(
 						sprintf(
 							'@Path "%s" in "%s::%s()" must starts with "/" (slash).',
@@ -59,7 +60,7 @@ class PathValidation implements IValidation
 				}
 
 				// MUST NOT: Ends with slash (/), except single '/' path
-				if (substr($path, -1, 1) === '/' && strlen($path) > 1) {
+				if (str_ends_with($path, '/') && strlen($path) > 1) {
 					throw (new InvalidSchemaException(
 						sprintf(
 							'@Path "%s" in "%s::%s()" must not ends with "/" (slash).',
@@ -112,6 +113,7 @@ class PathValidation implements IValidation
 				// -> -_
 				// @regex https://regex101.com/r/APckUJ/3
 				$matches = Regex::matchAll($path, '#\{(.+)\}#U');
+
 				if ($matches !== null) {
 					foreach ($matches as $item) {
 						$match = Regex::match($item[1], '#.*([^a-zA-Z0-9\-_]+).*#');
