@@ -15,17 +15,12 @@ use Doctrine\Common\Annotations\Annotation\Target;
 class Negotiation
 {
 
-	private string $suffix;
-
-	private bool $default;
-
-	private ?string $renderer;
-
-	public function __construct(string $suffix, bool $default = false, ?string $renderer = null)
+	public function __construct(
+		private readonly string $suffix,
+		private readonly bool $default = false,
+		private readonly ?string $renderer = null,
+	)
 	{
-		$this->suffix = $suffix;
-		$this->default = $default;
-		$this->renderer = $renderer;
 	}
 
 	public function getSuffix(): string
