@@ -15,20 +15,13 @@ use Doctrine\Common\Annotations\Annotation\Target;
 class RequestBody
 {
 
-	private ?string $description;
-
-	private ?string $entity;
-
-	private bool $required;
-
-	private bool $validation;
-
-	public function __construct(?string $description = null, ?string $entity = null, bool $required = false, bool $validation = true)
+	public function __construct(
+		private readonly ?string $description = null,
+		private readonly ?string $entity = null,
+		private readonly bool $required = false,
+		private readonly bool $validation = true,
+	)
 	{
-		$this->description = $description;
-		$this->entity = $entity;
-		$this->required = $required;
-		$this->validation = $validation;
 	}
 
 	public function getEntity(): ?string
