@@ -16,18 +16,14 @@ use Doctrine\Common\Annotations\AnnotationException;
 class Tag
 {
 
-	private string $name;
-
-	private ?string $value;
-
-	public function __construct(string $name, ?string $value = null)
+	public function __construct(
+		private readonly string $name,
+		private readonly ?string $value = null,
+	)
 	{
 		if ($name === '') {
 			throw new AnnotationException('Empty @Tag name given');
 		}
-
-		$this->name = $name;
-		$this->value = $value;
 	}
 
 	public function getName(): string
