@@ -9,11 +9,10 @@ use Apitte\Core\Mapping\RequestEntityMapping;
 class RequestEntityDecorator implements IRequestDecorator
 {
 
-	protected RequestEntityMapping $mapping;
-
-	public function __construct(RequestEntityMapping $mapping)
+	public function __construct(
+		protected RequestEntityMapping $mapping,
+	)
 	{
-		$this->mapping = $mapping;
 	}
 
 	public function decorateRequest(ApiRequest $request, ApiResponse $response): ApiRequest
