@@ -8,14 +8,11 @@ use Throwable;
 class DispatchError
 {
 
-	protected Throwable $error;
-
-	protected ApiRequest $request;
-
-	public function __construct(Throwable $error, ApiRequest $request)
+	public function __construct(
+		protected Throwable $error,
+		protected ApiRequest $request,
+	)
 	{
-		$this->error = $error;
-		$this->request = $request;
 	}
 
 	public function getError(): Throwable
