@@ -13,11 +13,10 @@ use Psr\Log\LogLevel;
 class PsrLogErrorHandler extends SimpleErrorHandler
 {
 
-	private LoggerInterface $logger;
-
-	public function __construct(LoggerInterface $logger)
+	public function __construct(
+		private readonly LoggerInterface $logger,
+	)
 	{
-		$this->logger = $logger;
 	}
 
 	public function handle(DispatchError $dispatchError): ApiResponse
