@@ -6,14 +6,11 @@ use Apitte\Core\Exception\Api\ValidationException;
 use Doctrine\Common\Annotations\AnnotationReader;
 use Doctrine\Common\Annotations\Reader;
 use Symfony\Component\Validator\ConstraintValidatorFactoryInterface;
-use Symfony\Component\Validator\ConstraintViolationListInterface;
 use Symfony\Component\Validator\Validation;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 class SymfonyValidator implements IEntityValidator
 {
-
-	private ?Reader $reader;
 
 	private ?ConstraintValidatorFactoryInterface $constraintValidatorFactory = null;
 
@@ -24,9 +21,10 @@ class SymfonyValidator implements IEntityValidator
 	/** @var list<string>|null */
 	private ?array $groups = null;
 
-	public function __construct(?Reader $reader = null)
+	public function __construct(
+		private ?Reader $reader = null,
+	)
 	{
-		$this->reader = $reader;
 		AnnotationReader::addGlobalIgnoredName('mapping');
 	}
 
@@ -77,11 +75,11 @@ class SymfonyValidator implements IEntityValidator
 
 		$validator = $validatorBuilder->getValidator();
 
-		/** @var ConstraintViolationListInterface $violations */
 		$violations = $validator->validate($entity, null, $this->groups);
 
 		if (count($violations) > 0) {
 			$fields = [];
+
 			foreach ($violations as $violation) {
 				$fields[$violation->getPropertyPath()][] = $violation->getMessage();
 			}
