@@ -10,11 +10,10 @@ use Nette\DI\Container;
 class RendererTransformer extends AbstractTransformer
 {
 
-	protected Container $container;
-
-	public function __construct(Container $container)
+	public function __construct(
+		protected Container $container,
+	)
 	{
-		$this->container = $container;
 	}
 
 	/**
