@@ -5,15 +5,13 @@ namespace Apitte\OpenApi\SchemaDefinition;
 class ArrayDefinition implements IDefinition
 {
 
-	/** @var mixed[] */
-	private array $data = [];
-
 	/**
 	 * @param mixed[] $data
 	 */
-	public function __construct(array $data)
+	public function __construct(
+		private readonly array $data,
+	)
 	{
-		$this->data = $data;
 	}
 
 	/**
