@@ -13,14 +13,11 @@ use Contributte\OpenApi\Utils\Helpers;
 class CoreDefinition implements IDefinition
 {
 
-	protected ApiSchema $schema;
-
-	private IEntityAdapter $entityAdapter;
-
-	public function __construct(ApiSchema $schema, IEntityAdapter $entityAdapter)
+	public function __construct(
+		protected ApiSchema $schema,
+		private readonly IEntityAdapter $entityAdapter,
+	)
 	{
-		$this->schema = $schema;
-		$this->entityAdapter = $entityAdapter;
 	}
 
 	/**
@@ -29,6 +26,7 @@ class CoreDefinition implements IDefinition
 	public function load(): array
 	{
 		$data = ['paths' => []];
+
 		foreach ($this->getEndpoints() as $endpoint) {
 			foreach ($endpoint->getMethods() as $method) {
 				$data['paths'][(string) $endpoint->getMask()][strtolower($method)] = $this->createOperation($endpoint);
@@ -49,6 +47,7 @@ class CoreDefinition implements IDefinition
 
 		// Tags
 		$tags = $this->getOperationTags($endpoint);
+
 		if ($tags !== []) {
 			$operation['tags'] = array_keys($tags);
 		}
@@ -59,6 +58,7 @@ class CoreDefinition implements IDefinition
 		}
 
 		$requestBody = $endpoint->getRequestBody();
+
 		if ($requestBody !== null) {
 			$operation['requestBody'] = $this->createRequestBody($requestBody);
 		}
@@ -68,9 +68,7 @@ class CoreDefinition implements IDefinition
 		// TODO deprecated
 		// $operation->setDeprecated(false);
 
-		$operation = Helpers::merge($endpoint->getOpenApi()['method'] ?? [], $operation);
-
-		return $operation;
+		return Helpers::merge($endpoint->getOpenApi()['method'] ?? [], $operation);
 	}
 
 	/**
@@ -85,11 +83,13 @@ class CoreDefinition implements IDefinition
 		}
 
 		$description = $requestBody->getDescription();
+
 		if ($description !== null) {
 			$requestBodyData['description'] = $description;
 		}
 
 		$entity = $requestBody->getEntity();
+
 		if ($entity !== null) {
 			$requestBodyData['content'] = [
 				// TODO resolve content types
@@ -109,6 +109,7 @@ class CoreDefinition implements IDefinition
 	protected function createResponses(Endpoint $endpoint): array
 	{
 		$responses = [];
+
 		foreach ($endpoint->getResponses() as $response) {
 			$responses[$response->getCode()] = $this->createResponse($response);
 		}
@@ -126,6 +127,7 @@ class CoreDefinition implements IDefinition
 		];
 
 		$entity = $response->getEntity();
+
 		if ($entity !== null) {
 			$responseData['content'] = [
 				// TODO resolve content types
@@ -150,6 +152,7 @@ class CoreDefinition implements IDefinition
 		];
 
 		$parameterDescription = $endpointParameter->getDescription();
+
 		if ($parameterDescription !== null) {
 			$parameter['description'] = $parameterDescription;
 		}
