@@ -20,7 +20,7 @@ final class AttributeMultiController extends ApiV1Controller
 	}
 
 	#[RequestParameter(name: 'name_value', type: 'type_value', in: 'path')]
-	#[RequestParameter(in: 'query', type: 'type_value_2', name: 'name_value_2')]
+	#[RequestParameter(name: 'name_value_2', type: 'type_value_2', in: 'query')]
 	public function requestParameters(): void
 	{
 		// Tests
